Move to SpawnItemsOnUseSystem
